requirements.txt
ollama==0.6.1
pydantic==2.12.5
srt==3.5.3
pysrt==2.12.5
google-generativeai
